Process ngrams, add syllable count, pronunciation.

+++ internal/dybmapi/model.go
@@ -7,7 +7,7 @@ type IndexData struct {
 type PimpedLine struct {
 	Number    int    `json:"number"`
 	Line      string `json:"line"`
-	Syllables int    `json:"syllables"`
+	Syllables byte   `json:"syllables"`
 }
 
 type PimpedLyrics struct {
